+++ src/main/java/de/maxhenkel/corpse/gui/DeathHistoryScreen.java
@@ -1,7 +1,7 @@
 package de.maxhenkel.corpse.gui;
 
 import com.mojang.authlib.GameProfile;
-import com.mojang.blaze3d.platform.GlStateManager;
+import com.mojang.blaze3d.systems.RenderSystem;
 import de.maxhenkel.corpse.Death;
 import de.maxhenkel.corpse.Main;
 import de.maxhenkel.corpse.net.MessageShowCorpseInventory;
@@ -14,7 +14,10 @@ import net.minecraft.entity.Pose;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.SoundEvents;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.text.*;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.TextComponentUtils;
+import net.minecraft.util.text.TextFormatting;
+import net.minecraft.util.text.TranslationTextComponent;
 import net.minecraft.util.text.event.ClickEvent;
 import net.minecraft.util.text.event.HoverEvent;
 
@@ -126,7 +129,7 @@ public class DeathHistoryScreen extends ScreenBase {
         drawRight(TextFormatting.GRAY + "" + Math.round(death.getPosZ()) + " Z", guiTop + 100);
 
         // Player
-        GlStateManager.color4f(1.0F, 1.0F, 1.0F, 1.0F);
+        RenderSystem.color4f(1.0F, 1.0F, 1.0F, 1.0F);
 
         RemoteClientPlayerEntity player = new RemoteClientPlayerEntity(minecraft.world, new GameProfile(death.getPlayerUUID(), death.getPlayerName())) {
             @Override
@@ -136,7 +139,7 @@ public class DeathHistoryScreen extends ScreenBase {
         };
         player.recalculateSize();
 
-        InventoryScreen.drawEntityOnScreen(guiLeft + xSize - (xSize - hSplit) / 2, guiTop + ySize / 2 + 30, 40, (guiLeft + xSize - (xSize - hSplit) / 2) - mouseX, (guiTop + ySize / 2) - mouseY, player);
+        InventoryScreen.func_228187_a_(guiLeft + xSize - (xSize - hSplit) / 2, guiTop + ySize / 2 + 30, 40, (guiLeft + xSize - (xSize - hSplit) / 2) - mouseX, (guiTop + ySize / 2) - mouseY, player);
 
         if (mouseX >= guiLeft + 7 && mouseX <= guiLeft + hSplit && mouseY >= guiTop + 70 && mouseY <= guiTop + 100 + font.FONT_HEIGHT) {
             renderTooltip(new TranslationTextComponent("tooltip.teleport").getFormattedText(), mouseX, mouseY);