+++ changelog.md
@@ -1,1 +1,1 @@
-- Added pl_pl language
+-